gooey_run now generates code for a script that runs the gui, the parser (aparser) is now in its own file
