+++ pysg/pysg_special_run.py
@@ -1,40 +1,15 @@
 import PySimpleGUI as sg
-from sys import argv
+from sys import argv, modules
 from re import match
 
-# parses the athinput file and returns a dictionary
-def parse(filename):
-    file = open(filename, 'r')
-    lines = file.readlines()
-    data = {}
-    info = {}
-    prefix = ''
-    # looking for name and abstract
-    # assuming name and abstract lines have no comments in them
-    for line in lines:
-        # this regex matches the section line:
-        # <[string]>
-        m = match('^\s*<(.+)>.*', line)
-        if m:
-            prefix = m.group(1)
-            continue
-        # this regex matches strings of the form:
-        # [string] = [string with spaces] # comment #> [string] [string]
-        m = match('^([^#]+)\s*=\s*([^#]+).*#>\s+([^\s]+)(\s+.+|\s*)$', line)
-        if m:
-            # strip the leading and trailing whitespace
-            # dictionary entry is a list
-            data[f'{prefix}_{m.group(1)}'.strip()] = [
-                m.group(2).strip(), 
-                m.group(3).strip(), 
-                m.group(4).strip()
-            ]
-            continue
-        # this regex matches the name / abstract
-        m = match('^([^#]+)\s*=\s*([^#]+).*$', line)
-        if m:
-            info[m.group(1).strip()] = m.group(2).strip()
-    return data, info
+# import aparser
+from os import environ
+from importlib.util import spec_from_file_location as sffl, module_from_spec as mfs
+spec = sffl('aparser', (environ['AGUI'] if 'AGUI' in environ else '~/agui') + '/aparser.py')
+aparser = mfs(spec)
+modules[spec.name] = aparser
+spec.loader.exec_module(aparser)
+parse = aparser.parse_special
 
 def build_layout():
     # for use with the dictionary
